Several bug fixes (#167)
* Checking for nil location in site test

* Avoiding nil pointer issues in export
Retrying WaitForState timeout in export

* Increasing sleeps

* Delete conflicting pools

* Adding configurable timeouts to resources with long cache TTLs

* Propagating correct error when reading user routing utilization

+++ genesyscloud/resource_genesyscloud_telephony_providers_edges_site_test.go
@@ -501,7 +501,7 @@ func deleteLocationWithNumber(emergencyNumber string) error {
 						return err
 					}
 					_, err = locationsAPI.DeleteLocation(*location.Id)
-					time.Sleep(10 * time.Second)
+					time.Sleep(8 * time.Second)
 					return err
 				}
 			}
@@ -525,12 +525,12 @@ func deleteSiteWithLocationId(locationId string) error {
 		}
 
 		for _, site := range *sites.Entities {
-			if *site.Location.Id == locationId {
+			if site.Location != nil && *site.Location.Id == locationId {
 				_, err := edgesAPI.DeleteTelephonyProvidersEdgesSite(*site.Id)
 				if err != nil {
 					return err
 				}
-				time.Sleep(10 * time.Second)
+				time.Sleep(8 * time.Second)
 				break
 			}
 		}
